refactor(reports): move SwitchRead out of profile
Keep the report read/unread toggle with report UI instead of the
profile directory.
